requirements.txt
langchain 
langchain-community
langchain_groq 
langchain-huggingface
langchain_experimental
langchain_core
langchain_huggingface
pyyaml
pinecone

